+++ domains/Users/Http/Resources/UserResource.php
@@ -19,6 +19,7 @@ class UserResource extends JsonResource
             'created_at' => $this->created_at,
             'updated_at' => $this->updated_at,
             'deleted_at' => $this->deleted_at,
+            'email_verified_at' => $this->email_verified_at,
             'customer' => CustomerResource::make($this->whenLoaded('customer')),
         ];
     }