+++ domains/Users/Tests/Feature/Controllers/UsersStoreControllerTest.php
@@ -4,9 +4,11 @@ namespace Domains\Users\Tests\Feature\Controllers;
 
 use Domains\Customers\Models\Customer;
 use Domains\Users\Models\User;
+use Illuminate\Auth\Notifications\VerifyEmail;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Foundation\Testing\WithFaker;
 use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Notification;
 use Laravel\Passport\Passport;
 use Tests\TestCase;
 
@@ -50,6 +52,8 @@ class UsersStoreControllerTest extends TestCase
     /** @test */
     public function system_operators_can_store_new_system_operators(): void
     {
+        Notification::fake();
+
         Passport::actingAs($this->systemOperator, [
             'organization:users:view',
             'organization:users:manage',
@@ -73,6 +77,9 @@ class UsersStoreControllerTest extends TestCase
 
         /** @var User $newUser */
         $newUser = User::email($payload['email'])->first();
+
+        Notification::assertSentTo($newUser, VerifyEmail::class);
+        $this->assertFalse($newUser->hasVerifiedEmail());
 
         $this->assertTrue(
             Hash::check($payload['password'], $newUser->password)
@@ -113,6 +120,8 @@ class UsersStoreControllerTest extends TestCase
     /** @test */
     public function system_operators_can_store_new_users_associated_with_any_customer(): void
     {
+        Notification::fake();
+
         $customer = factory(Customer::class)->create();
 
         Passport::actingAs($this->systemOperator, [
@@ -139,6 +148,9 @@ class UsersStoreControllerTest extends TestCase
         /** @var User $newUser */
         $newUser = User::email($payload['email'])->first();
 
+        Notification::assertSentTo($newUser, VerifyEmail::class);
+        $this->assertFalse($newUser->hasVerifiedEmail());
+
         $this->assertTrue(
             Hash::check($payload['password'], $newUser->password)
         );
@@ -149,16 +161,18 @@ class UsersStoreControllerTest extends TestCase
     /** @test */
     public function users_can_store_new_users(): void
     {
-        Passport::actingAs($this->user, [
-            'organization:users:view',
-            'organization:users:manage',
-        ]);
+        Notification::fake();
 
         $payload = [
             'name' => $this->faker->name,
             'email' => $this->faker->safeEmail,
             'password' => 'password',
         ];
+
+        Passport::actingAs($this->user, [
+            'organization:users:view',
+            'organization:users:manage',
+        ]);
 
         $this->postJson('/users', $payload)
             ->assertCreated();
@@ -172,6 +186,9 @@ class UsersStoreControllerTest extends TestCase
         /** @var User $newUser */
         $newUser = User::email($payload['email'])->first();
 
+        Notification::assertSentTo($newUser, VerifyEmail::class);
+        $this->assertFalse($newUser->hasVerifiedEmail());
+
         $this->assertTrue(
             Hash::check($payload['password'], $newUser->password)
         );
@@ -182,17 +199,17 @@ class UsersStoreControllerTest extends TestCase
     /** @test */
     public function it_fails_if_user_sends_customer_id_when_creating_other_users(): void
     {
-        Passport::actingAs($this->user, [
-            'organization:users:view',
-            'organization:users:manage',
-        ]);
-
         $payload = [
             'customer_id' => 1,
             'name' => $this->faker->name,
             'email' => $this->faker->safeEmail,
             'password' => 'password',
         ];
+
+        Passport::actingAs($this->user, [
+            'organization:users:view',
+            'organization:users:manage',
+        ]);
 
         $this->postJson('/users', $payload)
             ->assertForbidden();
@@ -220,16 +237,16 @@ class UsersStoreControllerTest extends TestCase
     /** @test */
     public function it_fails_if_email_is_already_registered_to_another_resource(): void
     {
-        Passport::actingAs($this->systemOperator, [
-            'organization:users:view',
-            'organization:users:manage',
-        ]);
-
         $payload = [
             'name' => $this->faker->name,
             'email' => $this->systemOperator->email,
             'password' => 'password',
         ];
+
+        Passport::actingAs($this->systemOperator, [
+            'organization:users:view',
+            'organization:users:manage',
+        ]);
 
         $this->postJson('/users', $payload)
             ->assertJsonValidationErrors(['email']);
@@ -238,17 +255,17 @@ class UsersStoreControllerTest extends TestCase
     /** @test */
     public function it_fails_if_vat_number_is_already_registered_to_another_resource(): void
     {
-        Passport::actingAs($this->systemOperator, [
-            'organization:users:view',
-            'organization:users:manage',
-        ]);
-
         $payload = [
             'name' => $this->faker->name,
             'email' => $this->faker->email,
             'vat_number' => $this->systemOperator->vat_number,
             'password' => 'password',
         ];
+
+        Passport::actingAs($this->systemOperator, [
+            'organization:users:view',
+            'organization:users:manage',
+        ]);
 
         $this->postJson('/users', $payload)
             ->assertJsonValidationErrors(['vat_number']);
@@ -257,17 +274,17 @@ class UsersStoreControllerTest extends TestCase
     /** @test */
     public function it_fails_if_customer_id_doesnt_exists(): void
     {
-        Passport::actingAs($this->systemOperator, [
-            'organization:users:view',
-            'organization:users:manage',
-        ]);
-
         $payload = [
             'name' => $this->faker->name,
             'email' => $this->faker->safeEmail,
             'password' => 'password',
             'customer_id' => 5,
         ];
+
+        Passport::actingAs($this->systemOperator, [
+            'organization:users:view',
+            'organization:users:manage',
+        ]);
 
         $this->postJson('/users', $payload)
             ->assertJsonValidationErrors(['customer_id']);