+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this repository will be documented in this file.
 ### Added
 
 - Users generated tokens are invalidated when syncing roles (#35)
+- A newly created User must verify their email before being able to login (#36)
 
 ### Changed